ai/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
distro==1.9.0
exceptiongroup==1.2.2
groq==0.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
pydantic==2.8.2
pydantic-core==2.20.1
python-dotenv==1.0.1
sniffio==1.3.1
typing-extensions==4.12.2
